+++ test/pow_assent/plug_test.exs
@@ -32,7 +32,8 @@ defmodule PowAssent.PlugTest do
     assert {:ok, url, conn} = Plug.authorize_url(conn, "test_provider", "https://example.com/")
 
     assert url =~ "http://localhost:8888/oauth/authorize?client_id=client_id&redirect_uri=https%3A%2F%2Fexample.com%2F&response_type=code&state="
-    assert Map.has_key?(conn.private, :pow_assent_state)
+    assert Map.has_key?(conn.private, :pow_assent_session_params)
+    assert conn.private[:pow_assent_session_params][:state]
   end
 
   describe "callback/3" do